+++ mayan/apps/platform/client_backends.py
@@ -4,6 +4,7 @@ import logging
 import sentry_sdk
 from sentry_sdk.integrations.celery import CeleryIntegration
 from sentry_sdk.integrations.django import DjangoIntegration
+from sentry_sdk.integrations.redis import RedisIntegration
 
 from django.conf.urls import url
 
@@ -39,7 +40,9 @@ class ClientBackendSentry(ClientBackend):
     def launch(self):
         kwargs = self.setup_arguments()
 
-        kwargs['integrations'] = (CeleryIntegration(), DjangoIntegration(),)
+        kwargs['integrations'] = (
+            CeleryIntegration(), DjangoIntegration(), RedisIntegration()
+        )
 
         logger.debug('cleaned arguments: %s', kwargs)